requirements.txt
aniso8601==9.0.1
blinker==1.8.2
boto3==1.34.158
botocore==1.34.158
cachetools==5.5.0
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==2.1.0
passlib==1.7.4
pillow==10.4.0
pycrypto==2.6.1
pycryptodome==3.20.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2024.1
s3transfer==0.10.2
setuptools==72.1.0
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
